Refactored AddSubmissionContent so it has a separate ViewModel for it's logic instead of the AssignmentDetailsViewModel.
