+++ libs/horizon/src/main/java/com/instructure/horizon/features/moduleitemsequence/content/assignment/AssignmentDetailsUiState.kt
@@ -15,10 +15,6 @@
  */
 package com.instructure.horizon.features.moduleitemsequence.content.assignment
 
-import android.net.Uri
-import androidx.annotation.StringRes
-import com.instructure.canvasapi2.models.Assignment
-import com.instructure.horizon.R
 import com.instructure.horizon.horizonui.organisms.cards.AttemptCardState
 import com.instructure.horizon.horizonui.platform.LoadingState
 
@@ -27,14 +23,14 @@ data class AssignmentDetailsUiState(
     val instructions: String = "",
     val ltiUrl: String = "",
     val submissionDetailsUiState: SubmissionDetailsUiState = SubmissionDetailsUiState(),
-    val addSubmissionUiState: AddSubmissionUiState = AddSubmissionUiState(),
     val showSubmissionDetails: Boolean = false,
     val showAddSubmission: Boolean = false,
     val toolsBottomSheetUiState: ToolsBottomSheetUiState = ToolsBottomSheetUiState(),
     val ltiButtonPressed: ((String) -> Unit)? = null,
     val urlToOpen: String? = null,
     val onUrlOpened: () -> Unit = {},
-    val submissionConfirmationUiState: SubmissionConfirmationUiState = SubmissionConfirmationUiState(),
+    val onSubmissionSuccess: suspend () -> Unit = {},
+    val submissionConfirmationUiState: SubmissionConfirmationUiState = SubmissionConfirmationUiState()
 )
 
 data class SubmissionDetailsUiState(
@@ -61,73 +57,6 @@ data class FileItem(
     val fileType: String,
     val thumbnailUrl: String,
     val fileId: Long
-)
-
-data class AddSubmissionUiState(
-    val submissionTypes: List<AddSubmissionTypeUiState> = emptyList(),
-    val selectedSubmissionTypeIndex: Int = 0,
-    val onSubmissionTypeSelected: (Int) -> Unit = {},
-    val onSubmissionButtonClicked: () -> Unit = {},
-    val showSubmissionConfirmation: Boolean = false,
-    val onDismissSubmissionConfirmation: () -> Unit = {},
-    val onSubmitAssignment: () -> Unit = {},
-    val submissionInProgress: Boolean = false,
-    val errorMessage: String? = null,
-)
-
-sealed class AddSubmissionTypeUiState(
-    @StringRes val labelRes: Int,
-    val submissionType: Assignment.SubmissionType,
-    open val draftUiState: DraftUiState = DraftUiState(),
-    open val submitEnabled: Boolean = false,
-) {
-
-    abstract fun copyWith(
-        draftUiState: DraftUiState = this.draftUiState,
-        submitEnabled: Boolean = this.submitEnabled
-    ): AddSubmissionTypeUiState
-
-    data class Text(
-        val text: String = "",
-        val onTextChanged: (String) -> Unit = {},
-        override val draftUiState: DraftUiState = DraftUiState(),
-        override val submitEnabled: Boolean = false
-    ) : AddSubmissionTypeUiState(R.string.assignmentDetilas_submissionTypeText, Assignment.SubmissionType.ONLINE_TEXT_ENTRY) {
-        override fun copyWith(
-            draftUiState: DraftUiState,
-            submitEnabled: Boolean
-        ) = copy(draftUiState = draftUiState, submitEnabled = submitEnabled)
-    }
-
-    data class File(
-        val allowedTypes: List<String> = emptyList(),
-        val cameraAllowed: Boolean = false,
-        val galleryPickerAllowed: Boolean = false,
-        val files: List<AddSubmissionFileUiState> = emptyList(),
-        val onFileAdded: (Uri) -> Unit = {},
-        override val draftUiState: DraftUiState = DraftUiState(),
-        override val submitEnabled: Boolean = false
-    ) : AddSubmissionTypeUiState(R.string.assignmentDetilas_submissionTypeFileUpload, Assignment.SubmissionType.ONLINE_UPLOAD) {
-
-        override fun copyWith(
-            draftUiState: DraftUiState,
-            submitEnabled: Boolean
-        ) = copy(draftUiState = draftUiState, submitEnabled = submitEnabled)
-    }
-}
-
-data class AddSubmissionFileUiState(
-    val name: String = "",
-    val path: String? = null,
-    val onDeleteClicked: () -> Unit = {},
-)
-
-data class DraftUiState(
-    val draftDateString: String = "",
-    val onDeleteDraftClicked: () -> Unit = {},
-    val showDeleteDraftConfirmation: Boolean = false,
-    val onDismissDeleteDraftConfirmation: () -> Unit = {},
-    val onDraftDeleted: () -> Unit = {},
 )
 
 data class ToolsBottomSheetUiState(
